requirements.txt
argopt==0.8.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==23.1.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
# channels==4.0.0
# channels-redis==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-environ==0.10.0
django-filter==23.2
django-jsonfield==1.4.1
django-leaflet==0.29.0
django-phonenumber-field==7.1.0
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.5
firebase-admin==6.2.0
geographiclib==2.0
geopy==2.3.0
git-fame==2.0.1
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
msgpack==1.0.5
numpy==1.25.0
packaging==23.1
paytm-pg==1.1.1
phonenumbers==8.13.13
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.23.4
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pyparsing==3.1.0
pyrsistent==0.19.3
pytz==2021.3
PyYAML==6.0
# redis==4.6.0
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==3.0.3
tabulate==0.9.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2021.5
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
